Stack swapping: Removed a printk.
The printk showed that new Windows threads are started with
stack swapping enabled. We have to disable it for a variety
of reasons. Since that Windows 'feature' comes from a time
where computers had 4MB (not GB) main memory we can afford
to disable stack swapping for all WinDRBD threads.